Tank tick tweaks
Previously, If a tank was completely emptied out the adjacent tanks
would not be notified to balance their contents with it. If an empty
tank was placed next to the bottom block of a column of multiple full
tanks, the full tank would not drain into the space opening up in the
tank below it.